+++ s3empty/logger.py
@@ -3,11 +3,11 @@
 from conflog import Conflog
 
 
-def init():
+def init(log_level: str) -> object:
     """Initialize logger."""
 
     cfl = Conflog(
-        conf_dict={"level": "info", "format": "[s3empty] %(levelname)s %(message)s"}
+        conf_dict={"level": log_level, "format": "[s3empty] %(levelname)s %(message)s"}
     )
 
     return cfl.get_logger(__name__)